Refactor pyth functions to their own module.
