+++ src/_pytask/path.py
@@ -1,0 +1,62 @@
+from pathlib import Path
+from typing import List
+
+
+def relative_to(path: Path, source: Path, include_source: bool = True):
+    """Make a path relative to another path.
+
+    In contrast to :meth:`pathlib.Path.relative_to`, this function allows to keep the
+    name of the source path.
+
+    Examples
+    --------
+    >>> from pathlib import Path
+    >>> relative_to(Path("folder", "file.py"), Path("folder")).as_posix()
+    'folder/file.py'
+    >>> relative_to(Path("folder", "file.py"), Path("folder"), False).as_posix()
+    'file.py'
+
+    """
+    return Path(source.name if include_source else "", path.relative_to(source))
+
+
+def find_closest_ancestor(path: Path, potential_ancestors: List[Path]):
+    """Find the closest ancestor of a path.
+
+    In case a path is the path to the task file itself, we return the path.
+
+    .. note::
+
+        The check :meth:`pathlib.Path.is_file` only succeeds when the file exists. This
+        must be true as otherwise an error is raised by :obj:`click` while using the
+        cli.
+
+    Examples
+    --------
+    >>> from pathlib import Path
+    >>> find_closest_ancestor(Path("folder", "file.py"), [Path("folder")]).as_posix()
+    'folder'
+
+    >>> paths = [Path("folder"), Path("folder", "subfolder")]
+    >>> find_closest_ancestor(Path("folder", "subfolder", "file.py"), paths).as_posix()
+    'folder/subfolder'
+
+    """
+    closest_ancestor = None
+    for ancestor in potential_ancestors:
+        if ancestor == path:
+            closest_ancestor = path
+            break
+
+        # Paths can also point to files in which case we want to take the parent folder.
+        if ancestor.is_file():
+            ancestor = ancestor.parent
+
+        if ancestor in path.parents:
+            if closest_ancestor is None or (
+                len(path.relative_to(ancestor).parts)
+                < len(path.relative_to(closest_ancestor).parts)
+            ):
+                closest_ancestor = ancestor
+
+    return closest_ancestor
